fix(suite): Show fiat value in AssetCard
